+++ dds_web/database/models.py
@@ -7,6 +7,7 @@
 # Standard library
 import datetime
 import os
+import time
 
 # Installed
 import sqlalchemy
@@ -16,6 +17,7 @@ import flask_login
 import pathlib
 from cryptography.hazmat.primitives.twofactor import (
     hotp as twofactor_hotp,
+    totp as twofactor_totp,
     InvalidToken as twofactor_InvalidToken,
 )
 from cryptography.hazmat.primitives import hashes
@@ -357,10 +359,16 @@ class User(flask_login.UserMixin, db.Model):
     username = db.Column(db.String(50), primary_key=True, autoincrement=False)
     name = db.Column(db.String(255), unique=False, nullable=True)
     _password_hash = db.Column(db.String(98), unique=False, nullable=False)
+    # 2fa columns
     hotp_secret = db.Column(db.LargeBinary(20), unique=False, nullable=False)
     hotp_counter = db.Column(db.BigInteger, unique=False, nullable=False, default=0)
     hotp_issue_time = db.Column(db.DateTime, unique=False, nullable=True)
+    totp_enabled = db.Column(db.Boolean, unique=False, nullable=False, default=False)
+    _totp_secret = db.Column(db.LargeBinary(64), unique=False, nullable=True)
+    totp_last_verified = db.Column(db.DateTime, unique=False, nullable=True)
+
     active = db.Column(db.Boolean, nullable=False, default=True)
+
     kd_salt = db.Column(db.LargeBinary(32), default=None)
     nonce = db.Column(db.LargeBinary(12), default=None)
     public_key = db.Column(db.LargeBinary(300), default=None)
@@ -489,6 +497,93 @@ class User(flask_login.UserMixin, db.Model):
         self.hotp_counter += 1
         # Reset the hotp_issue_time to allow a new code to be issued
         self.hotp_issue_time = None
+        db.session.commit()
+
+    @property
+    def totp_initiated(self):
+        """To check if activation of TOTP has been initiated, not to be confused
+        with user.totp_enabled, that indicates if TOTP is successfully enabled for the user."""
+        return self._totp_secret is not None
+
+    def totp_object(self):
+        """Google Authenticator seems to only be able to handle SHA1 and 6 digit codes"""
+        if self.totp_initiated:
+            return twofactor_totp.TOTP(self._totp_secret, 6, hashes.SHA1(), 30)
+        return None
+
+    def setup_totp_secret(self):
+        """Generate random 160 bit as the new totp secret and return provisioning URI
+        We're using SHA1 (Google Authenticator seems to only use SHA1 and 6 digit codes)
+        so secret should be at least 160 bits
+        https://cryptography.io/en/latest/hazmat/primitives/twofactor/#cryptography.hazmat.primitives.twofactor.totp.TOTP
+        """
+        self._totp_secret = os.urandom(20)
+        db.session.commit()
+
+    @property
+    def totp_secret_and_uri(self):
+        """Returns the users totp provisioning URI. Can only be sent before totp has been enabled."""
+        if self.totp_enabled:
+            # Can not be fetched again after it has been enabled
+            raise AuthenticationError("TOTP secret already enabled.")
+        totp = self.totp_object()
+
+        return self._totp_secret, totp.get_provisioning_uri(
+            account_name=self.username,
+            issuer="Data Delivery System",
+        )
+
+    def activate_totp(self):
+        """Set TOTP as the preferred means of second factor authentication.
+        Should be called after first totp token is verified
+        """
+        self.totp_enabled = True
+        db.session.commit()
+
+    def deactivate_totp(self):
+        """Fallback to HOTP as the preferred means of second factor authentication."""
+        self.totp_enabled = False
+        self._totp_secret = None
+        db.session.commit()
+
+    def verify_TOTP(self, token):
+        """Verify the totp token. Checks the previous, current and comming time frame
+        to allow for some clock drift.
+
+        raises AuthenticationError if token is invalid, has expired or
+        if totp has been successfully verified within the last 90 seconds.
+        """
+        # can't use totp successfully more than once within 90 seconds.
+        # Time frame chosen so that no one can use the same token more than once
+        # No need to use epoch time here.
+        current_time = dds_web.utils.current_time()
+        if self.totp_last_verified and (
+            current_time - self.totp_last_verified < datetime.timedelta(seconds=90)
+        ):
+            raise AuthenticationError(
+                "Authentications with time-based token need to be at least 90 seconds apart."
+            )
+
+        # construct object
+        totp = self.totp_object()
+
+        # attempt to verify the token using epoch time
+        # Allow for clock drift of 1 frame before or after
+        verified = False
+        for t_diff in [-30, 0, 30]:
+            verification_time = time.time() + t_diff
+            try:
+                totp.verify(token, verification_time)
+                verified = True
+                break
+            except twofactor_InvalidToken:
+                pass
+
+        if not verified:
+            raise AuthenticationError("Invalid time-based token.")
+
+        # if the token is valid, save time of last successful verification
+        self.totp_last_verified = current_time
         db.session.commit()
 
     # Email related