+++ dds_web/security/auth.py
@@ -128,6 +128,30 @@ def verify_password_reset_token(token):
     raise AuthenticationError(message="Invalid token")
 
 
+def verify_activate_totp_token(token, current_user):
+    claims = __verify_general_token(token)
+    user = __user_from_subject(claims.get("sub"))
+    if user and (user == current_user):
+        act = claims.get("act")
+        del claims
+        gc.collect()
+        if act and act == "totp":
+            return None
+    raise AuthenticationError(message="Invalid token")
+
+
+def verify_activate_hotp_token(token):
+    claims = __verify_general_token(token)
+    user = __user_from_subject(claims.get("sub"))
+    if user:
+        act = claims.get("act")
+        del claims
+        gc.collect()
+        if act and act == "hotp":
+            return user
+    raise AuthenticationError(message="Invalid token")
+
+
 def __base_verify_token_for_invite(token):
     """Verify token and return claims."""
     claims = __verify_general_token(token=token)
@@ -276,14 +300,15 @@ def __handle_multi_factor_authentication(user, mfa_auth_time_string):
             if mfa_auth_time >= dds_web.utils.current_time() - MFA_EXPIRES_IN:
                 return user
 
-        send_hotp_email(user)
+        error_message = ""
+        if not user.totp_enabled:
+            send_hotp_email(user)
+            error_message = "Please check your primary e-mail!"
 
         if flask.request.path.endswith("/user/second_factor"):
             return user
 
-        raise AuthenticationError(
-            message="Two-factor authentication is required! Please check your primary e-mail!"
-        )
+        raise AuthenticationError(message=f"Two-factor authentication is required! {error_message}")
 
 
 def send_hotp_email(user):
@@ -362,5 +387,6 @@ def verify_password(username, password):
     user = models.User.query.get(username)
 
     if user and user.is_active and user.verify_password(input_password=password):
-        send_hotp_email(user)
+        if not user.totp_enabled:
+            send_hotp_email(user)
         return user